Rename package symboltable to semanticAnalysis
