+++ dsl/src/interpreter/DSLInterpreter.java
@@ -15,9 +15,10 @@ import parser.AST.*;
 // CHECKSTYLE:ON: AvoidStarImport
 import parser.DungeonASTConverter;
 import runtime.MemorySpace;
+import runtime.Value;
 // importing all required classes from symbolTable will be to verbose
 // CHECKSTYLE:OFF: AvoidStarImport
-import runtime.Value;
+import semanticAnalysis.*;
 // CHECKSTYLE:ON: AvoidStarImport
 import symboltable.*;
