+++ dsl/src/runtime/Value.java
@@ -1,14 +1,14 @@
 package runtime;
 
-import symboltable.IType;
+import semanticAnalysis.IType;
 
 // TODO: should this be able to be undefined?
 
 /**
  * This class is used to represent a value in a {@link MemorySpace}, that is a combination of actual
- * value, dataType (defined by {@link IType} and optional reference to a {@link symboltable.Symbol}
- * from a {@link symboltable.SymbolTable}, as this class is basically the runtime equivalent of a
- * {@link symboltable.Symbol}
+ * value, dataType (defined by {@link IType} and optional reference to a {@link
+ * semanticAnalysis.Symbol} from a {@link semanticAnalysis.SymbolTable}, as this class is basically
+ * the runtime equivalent of a {@link semanticAnalysis.Symbol}
  */
 public class Value {
     public static Value NONE = new Value(null, null, -1, false);
@@ -37,7 +37,7 @@ public class Value {
     }
 
     /**
-     * Getter for index of the {@link symboltable.Symbol} linked to this Value
+     * Getter for index of the {@link semanticAnalysis.Symbol} linked to this Value
      *
      * @return index of the linked Symbol
      */
@@ -64,7 +64,7 @@ public class Value {
      *
      * @param dataType The datatype of this value
      * @param internalValue The actual value stored in this value
-     * @param symbolIdx The index of the {@link symboltable.Symbol} this Value corresponds to
+     * @param symbolIdx The index of the {@link semanticAnalysis.Symbol} this Value corresponds to
      */
     public Value(IType dataType, Object internalValue, int symbolIdx) {
         this.value = internalValue;
