+++ dsl/src/runtime/nativeFunctions/NativePrint.java
@@ -3,11 +3,11 @@ package runtime.nativeFunctions;
 import interpreter.DSLInterpreter;
 import java.util.List;
 import parser.AST.Node;
-import symboltable.BuiltInType;
-import symboltable.ICallable;
-import symboltable.IScope;
-import symboltable.ScopedSymbol;
-import symboltable.Symbol;
+import semanticAnalysis.BuiltInType;
+import semanticAnalysis.ICallable;
+import semanticAnalysis.IScope;
+import semanticAnalysis.ScopedSymbol;
+import semanticAnalysis.Symbol;
 
 // TODO: how to enable semantic analysis for this? e.g. parameter-count, etc.
 public class NativePrint extends ScopedSymbol implements ICallable {
