+++ dsl/src/semanticAnalysis/BuiltInType.java
@@ -1,4 +1,4 @@
-package symboltable;
+package semanticAnalysis;
 
 public class BuiltInType extends Symbol implements IType {
     public static BuiltInType intType = new BuiltInType("int", Scope.NULL);
