+++ dsl/src/semanticAnalysis/FunctionCallResolver.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package symboltable;
+package semanticAnalysis;
 
 import parser.AST.AstVisitor;
 import parser.AST.FuncCallNode;
