+++ dsl/src/semanticAnalysis/Symbol.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package symboltable;
+package semanticAnalysis;
 
 /** Represents a symbol in a program */
 public class Symbol {
