+++ dsl/src/semanticAnalysis/typebulder/DSLTypeMember.java
@@ -1,0 +1,2 @@
+package semanticAnalysis.typebulder;public @interface DSLTypeMember {
+}
