+++ game/src/dslToGame/TestComponent.java
@@ -1,0 +1,2 @@
+package dslToGame;public class TestComponent {
+}
